refactor(hush-cli): extract auth from pkg_cli
Move CallerAuthHeaders + build_caller_auth_headers +
add_trusted_publisher_signed_payload + private encode_signed_field
into pkg_cli/auth.rs. Struct, its fields, and the two consumed fns
widened to pub(super) for cross-module access by the registry-mutating
handlers; encode_signed_field stays private.
